feat: Megrating to the reducer protocol
