+++ Alfheim/Shared/Features/Account/Edit/EditAccountView.swift
@@ -1,6 +1,6 @@
 //
-//  AccountComposer.swift
-//  AccountComposer
+//  EditAccountView.swift
+//  Alfheim
 //
 //  Created by alex.huo on 2021/8/15.
 //  Copyright © 2021 blessingsoft. All rights reserved.
@@ -9,21 +9,21 @@
 import SwiftUI
 import ComposableArchitecture
 
-struct AccountComposer: View {
+struct EditAccountView: View {
   @Environment(\.dismiss) var dismiss
-  let store: Store<AccountEdit.State, AccountEdit.Action>
+  let store: Store<EditAccount.State, EditAccount.Action>
 
-  let mode: AccountEditor.Mode
+  let mode: EditAccount.Mode
 
   var body: some View {
     WithViewStore(store) { vs in
       NavigationView {
-        AccountEditor(store: store)
+        EditAccountForm(store: store)
           .navigationTitle(vs.isNew ? "New Account" : "Edit Account")
           .toolbar {
             ToolbarItem(placement: .confirmationAction) {
               Button {
-                let action = AccountEdit.Action.save(vs.snapshot, mode: vs.isNew ? .new : .update)
+                let action = EditAccount.Action.save(vs.snapshot, mode: vs.isNew ? .new : .update)
                  vs.send(action)
                 dismiss()
               } label: {
