+++ Alfheim/Shared/Features/Main/MainView.swift
@@ -90,7 +90,7 @@ struct ListNavigation: View {
   let store: Store<App.State, App.Action>
 
   var body: some View {
-    NavigationView {
+    NavigationStack {
       WithViewStore(store.stateless) { viewStore in
         ContentView(store: store)
           .task {
@@ -134,10 +134,10 @@ struct ContentView: View {
           }
         }
         .sheet(isPresented: vs.binding(get: \.isAccountComposerPresented, send: App.Action.addAccount)) {
-          AccountComposer(
+          EditAccountView(
             store: store.scope(
-              state: \.accountEditor,
-              action: App.Action.accountEditor),
+              state: \.editAccount,
+              action: App.Action.editAccount),
             mode: .new
           )
         }
@@ -155,248 +155,12 @@ struct ContentView: View {
   }
 }
 
-struct HomeView: View {
-  let store: Store<App.State, App.Action>
-
-  var body: some View {
-    WithViewStore(store.scope(state: \.homeState)) { vs in
-      List {
-        Section {
-          GridMenu(store: store)
-            .listRowBackground(Color(.systemGroupedBackground))
-            .buttonStyle(.plain)
-            .onTapGesture {}
-            .onLongPressGesture {}
-        }
-        .listRowSeparator(.hidden)
-        .listRowInsets(EdgeInsets())
-        .listRowBackground(Color(.systemGroupedBackground))
-        .buttonStyle(.plain)
-
-        Section {
-          OutlineGroup(vs.rootAccounts, children: \.children) { account in
-            AccountRow(store: store, account: account)
-              .contextMenu(account.root ? nil : ContextMenu {
-                Button {
-                  vs.send(.editAccount(presenting: true, account))
-                } label: {
-                  Label("Edit", systemImage: "pencil.circle")
-                }
-
-                Button(role: .destructive) {
-                  vs.send(.deleteAccount(account))
-                } label: {
-                  Label("Delete", systemImage: "trash.circle")
-                    .foregroundColor(.red)
-                }
-              })
-          }
-        } header: {
-          Text("Accounts").font(.headline).foregroundColor(.primary)
-        }
-        .listRowInsets(EdgeInsets(top: 4, leading: 12, bottom: 4, trailing: 16))
-        .sheet(isPresented: vs.binding(get: \.isEditingAccount, send: { App.Action.editAccount(presenting: $0, nil) })) {
-          AccountComposer(
-            store: store.scope(
-              state: \.accountEditor,
-              action: App.Action.accountEditor),
-            mode: .new
-          )
-        }
-      }
-      .buttonStyle(.plain)
-    }
-  }
-}
-
-/// Don't use AccountList in Home
-/// children 更新时，不会刷新！
-//struct AccountList: View {
-//  let store: Store<AppState, AppAction>
-//  let accounts: [Account]
-//
-//  var body: some View {
-//    List(accounts, children: \.children) { account in
-//      AccountRow(store: store, account: account)
-//    }
-//  }
-//}
-
-private struct AccountRow: View {
-  let store: Store<App.State, App.Action>
-  let account: Domain.Account
-
-  var body: some View {
-    ZStack {
-      HStack {
-        Text("\(account.emoji ?? "") \(account.name)")
-        Spacer()
-      }
-      WithViewStore(store.scope(state: \.rowState)) { vs in
-        NavigationLink(
-          tag: account.id,
-          selection: vs.binding(
-            get: \.selectionID,
-            send: App.Action.selectAccount(id:)
-          )
-        ) {
-          IfLetStore(
-            store.scope(
-              state: \.selection?.value,
-              action: App.Action.overview
-            ),
-            then: OverviewView.init(store:)
-          )
-        } label: {
-          EmptyView()
-        }
-        .buttonStyle(.plain)
-        .opacity(account.hasChildren ? 0: 1)
-      }
-    }
-  }
-}
-
-struct QuickMenu: View {
-  let store: Store<App.State, App.Action>
-  @State private var selection: Int? = nil
-  private var columns: [GridItem] = Array(repeating: .init(.flexible(), spacing: 18), count: 2)
-
-  init(store: Store<App.State, App.Action>) {
-    self.store = store
-    self.columns = Array(repeating: .init(.flexible(), spacing: 18), count: 2)
-  }
-
-  var body: some View {
-    WithViewStore(store.scope(state: \.sidebar)) { vs in
-      LazyVGrid(columns: columns, spacing: 18) {
-        ForEach(vs.menus) { item in
-          Button {
-            selection = item.id
-          } label: {
-            MenuRow(item: item, isSelected: selection == item.id)
-          }
-          .background(
-            NavigationLink(tag: item.id, selection: $selection, destination: {
-              Text(item.name)
-            }, label: {
-              EmptyView()
-            })
-          )
-        }
-      }
-      .onLongPressGesture {}
-    }
-  }
-}
-
-struct GridMenu: View {
-  @Environment(\.managedObjectContext) var viewContext // FIXME: use store environment
-  let store: Store<App.State, App.Action>
-  private var columns: [GridItem] = Array(repeating: .init(.flexible(), spacing: 18), count: 2)
-
-  init(store: Store<App.State, App.Action>) {
-    self.store = store
-    self.columns = Array(repeating: .init(.flexible(), spacing: 18), count: 2)
-  }
-
-  var body: some View {
-    WithViewStore(store.scope(state: \.sidebar)) { vs in
-      LazyVGrid(columns: columns, spacing: 18) {
-        ForEach(vs.menus) { item in
-          NavigationRow(
-            tag: item.id,
-            selection: vs.binding(get: \.selection?.id, send: { App.Action.selectMenu(selection: $0) })) {
-              IfLetStore(
-               store.scope(
-                state: \.sidebar.selection?.value,
-                action: App.Action.transaction
-               ),
-               then: TransactionList.init(store:)
-              )
-          } label: {
-            MenuRow(item: item, isSelected: vs.selection?.id == item.id)
-              .onTapGesture {
-                vs.send(.selectMenu(selection: item.id))
-              }
-          }
-        }
-      }
-      .onLongPressGesture {}
-      .onAppear {
-        DispatchQueue.main.asyncAfter(deadline: .now() + .milliseconds(400)) {
-          // selection = nil
-          vs.send(.selectMenu(selection: nil))
-        }
-      }
-    }
-  }
-}
-
-struct MenuRow: View {
-  let item: App.State.Sidebar.MenuItem
-  let isSelected: Bool
-
-  var body: some View {
-    VStack(alignment: .leading, spacing: 6) {
-      HStack {
-        Image(systemName: item.symbol).font(.title2).foregroundColor(item.color)
-        Spacer()
-        Text(item.value).font(.subheadline)
-      }
-      Text(item.name).font(.callout).fontWeight(.medium)
-    }
-    .padding(12)
-    .background(Color(UIColor.systemGray4).opacity(isSelected ? 1.0 : 0.0))
-    .background(Color(UIColor.secondarySystemGroupedBackground))
-    .clipShape(RoundedRectangle(cornerRadius: 18, style: .continuous))
-  }
-}
-
-extension EdgeInsets {
-  static let `default` = EdgeInsets(top: 4, leading: 12, bottom: 4, trailing: 16)
-}
-
-extension App.State.Sidebar.MenuItem {
-  var color: Color {
-    switch filter {
-    case .all: return Color.red
-    case .uncleared: return Color.gray
-    case .repeating: return Color.yellow
-    case .flagged: return Color.blue
-    }
-  }
-}
-
 #if DEBUG
-//struct AccountMenu_Previews: PreviewProvider {
-//  static var previews: some View {
-//    VStack {
-//      AccountMenu()
-//    }
-//    .frame(maxWidth: .infinity, maxHeight: .infinity)
-//    .background(.secondary)
-//  }
-//}
 
-struct Home_Previews: PreviewProvider {
+struct MainView_Previews: PreviewProvider {
   static var previews: some View {
-    List {
-      Section(header: EmptyView()) {
-        QuickMenu(store: AppStore(initialState: App.State(), reducer: RealWorld())).listRowBackground(Color.clear)
-      }
-      .listRowSeparator(.hidden)
-      .listRowInsets(EdgeInsets())
-
-      Section(header: Text("Accounts").font(.headline).foregroundColor(.primary)) {
-        Text("ABC")
-      }
-    }.listStyle(.insetGrouped)
+    MainView(store: AppStore(initialState: App.State(), reducer: RealWorld()))
   }
 }
-//struct MainView_Previews: PreviewProvider {
-//  static var previews: some View {
-//    MainView().environmentObject(AppStore(moc: viewContext))
-//  }
-//}
+
 #endif
