+++ Alfheim/Shared/Features/Main/Sidebar.swift
@@ -1,0 +1,9 @@
+//
+//  Sidebar.swift
+//  Alfheim
+//
+//  Created by alex.huo on 2022/12/5.
+//  Copyright © 2022 blessingsoft. All rights reserved.
+//
+
+import Foundation
